Fix Side+Navbar

+++ server/src/profile/profile.controller.ts
@@ -6,6 +6,8 @@ import {
   Param,
   Delete,
   Put,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { ProfileService } from './profile.service';
 
@@ -20,20 +22,32 @@ export class ProfileController {
     private profileService: ProfileService,
     private userService: UserService,
   ) { }
-
   @Post()
   async create(@Body() createProfileDto: CreateProfileDto): Promise<Profile> {
-    const newProfile = await this.profileService.create(createProfileDto);
-    if (!newProfile) {
-      try {
-        await this.userService.remove(createProfileDto.id);
-      } catch (error) {
-        throw new Error(error);
+    try {
+      const isExist = await this.profileService.findOne(createProfileDto.id);
+      if (isExist) {
+        throw new HttpException(
+          'Profile already exists',
+          HttpStatus.BAD_REQUEST,
+        );
       }
-    } else {
-      this.userService.update(createProfileDto.id, { profile: newProfile.id });
+      const newProfile = await this.profileService.create(createProfileDto);
+      if (!newProfile) {
+        try {
+          await this.userService.remove(createProfileDto.id);
+        } catch (error) {
+          throw new Error(error);
+        }
+      } else {
+        this.userService.update(createProfileDto.id, {
+          profile: newProfile.id,
+        });
+      }
+      return newProfile;
+    } catch (error) {
+      throw error;
     }
-    return newProfile;
   }
 
   @Get()
